Update query hook (#208)
* Retain Statement when parsing

When we parse the sql, retain the parsed statement as well as the raw
string and the plan.  This is because we have upcoming cases where we
want to reference the statement without needing to re-parse the query.

* Add extended query support to QueryHook

Add extended query support to QueryHook,.

Also re-use the parsed statement instead of parsing it again because it
is was wasted work, and it also made the code less clean.

(We still potentially handle the params more than once, which can be
fixed later)

Finally, ensure the client is Send + Sync. (Should this be enforced in
pgwire?)

* Add handle_extended_parse_query to QueryHook

Add handle_extended_parse_query to QueryHook so that parse phase
operations can be processed.

* Apply suggestion from @sunng87

---------

Co-authored-by: Ning Sun <classicning@gmail.com>